+++ ext/FlexiChainsMakieExt/traceplot.jl
@@ -15,7 +15,8 @@ function FC.mtraceplot(
         legend = (;),
         kwargs...,
     )
-    keys_to_plot = FC.PlotUtils.get_keys_to_plot(chn, param_or_params)
+    chn = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
+    keys_to_plot = keys(chn)
     isempty(keys_to_plot) && throw(ArgumentError("no parameters to plot"))
     nrows, ncols = if isnothing(layout)
         length(keys_to_plot), 1
@@ -48,13 +49,15 @@ This handles plotting onto a single Axis.
 function FC.mtraceplot(grid::MakieGrids, chn::FC.FlexiChain, param; axis = (;), kwargs...)
     # TODO: Error if there is already something at the grid position?
     # See e.g. https://github.com/rafaqz/DimensionalData.jl/blob/6db30de4b2e1fc7f8611b7e1dc3f89dc02c78598/ext/DimensionalDataMakieExt.jl#L85-L96
-    k = only(FC.PlotUtils.get_keys_to_plot(chn, param))
+    chn = FC.PlotUtils.subset_and_split_chain(chn, param)
+    k = only(keys(chn))
     return FC.mtraceplot!(
         Makie.Axis(grid; _default_traceplot_axis(k)..., axis...), chn, param; kwargs...
     )
 end
 function FC.mtraceplot!(ax::Makie.Axis, chn::FC.FlexiChain, param; kwargs...)
-    k = only(FC.PlotUtils.get_keys_to_plot(chn, param))
+    chn = FC.PlotUtils.subset_and_split_chain(chn, param)
+    k = only(keys(chn))
     a, p = FC.mtraceplot!(ax, FC.PlotUtils.FlexiChainTrace(chn, k); kwargs...)
     return Makie.AxisPlot(a, p)
 end